Refactor- move process events list data inside lib folder
